refactor: Change start_time type to double and improve logging error handling

+++ includes/CgiRequestHandler.hpp
@@ -6,7 +6,7 @@
 /*   By: ebmarque <ebmarque@student.42porto.com     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/09 15:34:18 by ebmarque          #+#    #+#             */
-/*   Updated: 2024/09/21 17:23:10 by ebmarque         ###   ########.fr       */
+/*   Updated: 2024/09/23 22:07:14 by ebmarque         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ typedef struct pollfd t_pollfd;
 typedef struct _s_client_process
 {
 	std::string method;
-	long start_time;
+	double start_time;
 	int client_fd;
 	int cgi_fd;
 } t_client_process;
